Move heartbeats from Scheduler to Worker
Previously the scheduler would initiate a connection to each worker-host
asking for health status.

Now each worker process (possibly several per host) reach out to the
scheduler to inform it of both their health and the fact that they're
around.  It does this on the normal "add_worker" channel, which has
become idempotent.  This has a few effects:

1.  If the scheduler gets restarted it eventually learns of the workers
    around it from heartbeats (which are just add_worker messages)
2.  The workers can include per-process information in the heartbeat
    rather than just per-host information
3.  There are possibly many more heartbeat connections (bad)