Track suspicious tasks and kill after enough failures
Fixes #301

This adds new state, Schduler.suspicious_tasks that counts the number
of times that a task was running while a worker quit.  Any task that
is involved in more than three workers failing is marked as bad and
sent back to the client.

This avoids a cascade of failures throughout the cluster.