Clear futures after restart
This clears out the references to all known futures from the executor
on a successful restart.  To avoid future reference decrements when the
futures do eventually garbage collect we add a `generation` attribute
to the executor to track how many restarts have occurred.  Reference
counts are only decremented if the generation of the Future matches
the current generation of the Executor.