Start Diagnostics server by default
Previously when calling `Executor()` or `LocalCluster()` we wouldn't
start up a bokeh server by default, instead asking users to find
`e.cluster.start_diagnostics_server()`.  Now we start this by default.

Additionally, we used to pass information via a `.dask-web-ui.json` file
which was written to the local directory, but searched for in the
install directory.  Now we write this to a `~/.dask/` directory instead
for global access.

Finally we turn off bokeh logging to clean up the interactive experience
and remove some unused bokeh plots (tables) that were using deprecated
functionality and causing warnings to pop up.

Fixes #402