Fix when moving to processing just before worker fails
Transition_ready_processing sends a task to a worker.  This can fail
if the worker dies just at the right time.  We've correctly been
calling remove_worker in this context, however we did this in the
middle of the transition, which was an unsafe state.  Now we do
it at the end