Add coroutine to synchronize worker data
It's possible that some extra data is getting lost on the workers.
This adds a scheduler coroutine that does a proper synchronization with
the worker.

It's not clear yet whether or not this is necessary.  If it is
necessary then we can put it in a long running PeriodicCallback.

Fixes https://github.com/dask/distributed/issues/333