Test dropped messages
I've gotten a report that dask.distributed has dropped a message.
This happened when many workers requested data from each other,
possibly while also running GIL-holding operations.

We can resolve this by taking away the persistent worker-worker
connections, but these are nice for performance and I'd also like
to get to the bottom of the issue.

This commit starts a stress test.  So far it's passing.