Launch tasks from other tasks
Modified ThreadPoolExecutor to support threads leaving the thread pool

This includes a global `secede` method that a submitted function can call to
have its thread leave the ThreadPoolExecutor's thread pool.  This allows the
thread pool to allocate another thread if necessary and so is useful when a
function realises that it is going to be a long-running job that doesn't want
to take up space.  When the function finishes its thread will terminate
gracefully.

use threadpoolexecutor in worker

Add new WorkerExecutor class

This is the client that a Worker gets internally.  Currently it is
exactly the same, except that `_scatter` dumps to the local worker's
data.

Remove deadlock from long-running tasks

Previously when tasks that created clients ran, they might a slot on the
scheduler for a long time, potentially causing deadlocks.  Now we
increase the number of slots for that worker whenever a task spins up a
client, removing the deadlock but opening ourselves up to
over-saturation.

update docs

add WorkerExecutor._gather method

add docstrings

add PSF license