Add Spill-to-disk policy for workers
This enacts a few changes:

First, workers can spill excess data to disk.  They use the core.dumps system
for writing data (usually just pickle) and then compress with the default
compression.  This uses ``zict.Buffer``, which uses an LRU mechanism to dump
excess data to a directory of files.

By default this is off.  You can turn it on by setting `--spill-bytes=auto` in
dask-worker or `--spill-bytes=1e9` or whatever to specify that it should spill
after using up one gigabyte.

Currently this uses `sizeof` to determine the amount of RAM that data is taking
up.  This is error prone with custom objects but fine as long as most things
are numpy arrays or pandas dataframes or whatnot.

Excess data is kept in the worker directory, which can now grow to be very
large.  We appropriately are now much more careful about cleaning this up when
workers close down.

We pass more information to the diagnostics so that disk-loads show up in the
task stream plot.  They show up as a slightly different shade of red than
worker-worker data transfers.

Fixes https://github.com/dask/distributed/issues/475