Clean up rpc objects explicitly
Previously we depended on `__del__` and garbage collection to clean up
rpc objects.  This caused some difficulties, particularly on Windows.

Now we clean up all rpc objects explicitly.  Most of this PR commit is
just tweaking other tests to be more robust.

We test this by explicitly checking the rpc count in the gen_cluster decorator
and the cluster testing context manager.