Allow any number of transfer logs per task
Previously a task-finished message from the worker to the scheduler could
contain three timings:

1.  compute_start/compute_stop
2.  transfer_start/transfer_stop
3.  disk_load_start/disk_load_stop

These would appear on the task stream plot for visual diagnostics.

This was fine, except that some tasks had multiple transfers, and so we were
only seeing the most recent one.

Now we include a list of arbitrary events that have occurred in support of any
particular task.  This can include multiple transfer events or even other
events that we haven't considered here, like serialization cost or some other
hardware cost that might come up in the future.