Use Connection Pool for Inter-worker communication
Previously we made new connections any time any worker wanted to request data
from another worker.  Now we use the pre-existing connection pool for this
purpose.  This lets us avoid reconnection costs.

We had this feature before but removed it due to errors that we couldn't track
down.  Given how much has changed sense them I'm comfortable adding this back
in, but we should be aware of the potential problem.