Add `scatter` option to joblib backend
Allows to prescatter data to the cluster. For large arguments (big
arrays, etc...) that are used in more than one task, this can be more
efficient as it avoids serializing the data for every task.

Also fix to ensure distributed backend is an instance of the
`ParallelBackendBase` for both scikit-learn and joblib modules.