Close things to avoid logging errors
Previously closing down a LocalCluster would result in a number of
errors being emitted to the console.  Now we close things in a different
order and check closing status to avoid these errors.

I tried to test this for a long while but failed to capture logging
information in pytest.