Use normal Client API in asynchronous mode
Previously we had two parallel APIs for synchronous/blocking and
asynchronous/non-blocking operation with underscored functions being
asynchronous:

    client.gather(...)  # synchronous
    yield/await client._gather(...)  # asynchronous

Now that we ask the user to explicitly provide an asychronous=True flag
we can make this choice for them and use the same function name in both cases.

    client = yield Client(asynchronous=True)
    yield/await client.gather(...)

This was a common mistake among new users who found their programs hanging.
This also allows us to have a single set of public facing functions.