Inform client of task state when we create new Futures
This is relevant when a new future gets a Future from a Queue or Variable.
We would like them to know the state of that queue or variable immediately
rather than have to wait.

This could be improved in the future by also sending along type or
exception/traceback information.  As it stands currently they still have to
wait a moment for this information to come down through normal channels.