Allow recreate on persisted collections
Previously only worked for single futures (which could have been pointing
to a compute of a collection), now works for persisted collections.

Suggestions for how to test for a non-persisted collection and improve
the error message?

Suggestions on whether there should be an action taken if local recreation
doesn't raise anything (this has happened to me over credentials).