Don't send closing messages on closed comm
When client closes it sends out a bunch of final messages to the scheduler
cleaning up its various resources.  However if the scheduler connection is
already dead then this generates annoying error messages.  This was a common
occurrence when closing a LocalCluster.

Reported in https://github.com/dask/distributed/issues/1294