Exit workers immediately if parent process exits
This is my response to @mrocklin in
[#1322](https://github.com/dask/distributed/issues/1322#issuecomment-322317319).

---

Prior to this commit, if the parent process was killed (for example due
to out-of-memory 'OOM'), then the child processes would never exit.

This could result in (potentially large) resources being tied up until
they are manually released, for example by an operator finding and killing
the processes.

This commit introduces a mechanism that enables children to determine
that the parent went away, and does not rely on the parent having time
to send any signals - so it works if the parent is SIGKILL'd.

The mechanism is described in detail in the code comments in this patch.

The test program I use to see the behaviour before and after this
commit. Without this commit, 1GiB of memory is leaked, and must be freed
with something like `pkill -f forkserver` (where `forkserver` uniquely
matches the command-line of what you intend to kill, which `forkserver`
does for me).

```python
import os
import signal
import sys

import distributed
from dask import delayed

def leak_memory():
    # Leak approx 1GiB of RAM (8 bytes per slot)
    sys.leak = [None for i in range(int(2**30 / 8))]
    return 42

if __name__ == "__main__":
    c = distributed.Client()
    q = delayed(leak_memory)()
    ret = q.compute()
    # Kill self. This results in the children remaining around since
    # they aren't signalled that we are done with them.
    os.kill(os.getpid(), signal.SIGKILL)
```

Possible fix for #1322.
/cc @pitrou @mrocklin