Add scheduler HTML routes to bokeh server
This adds a new set of routes to the HTTP server that are focused on internal
operation within the scheduler.  These template generated HTML.  They do not
update like the Bokeh plots but they are much more informative and highly
networked.

This needs additional exposure to more scheduler state
This needs some CSS love