Avoid double-scheduling in work stealing
This avoids double-scheduling tasks when work stealing by first checking
with the victim worker to verify that it has not yet started execution.
If it has started then we reverse the steal operation, if it has not
started then we cleanly start the new task only after the victim worker
has removed the task.

This passes most tests (those tests within stealing.py that if fails are
reasonable failures) but typically this sort of change triggers some
knock-on effects.