Support work stealing for resource restricted tasks (see issue #1389)
Idea is to naturally extend the current stealing approach: If a key has resource restrictions, choose the next idle worker providing the required resources rather than simply the next idle worker.

Modifications in stealing.py:
- find next worker providing the necessary resources
- make sure resources are properly consumed and released during stealing

Added tests to verify that
- resource restricted tasks are stolen when workers with sufficient resources are available (and resources are properly consumed and released)
- resource restricted tasks are not stolen when available workers do not possess sufficient resources

Mentioned the support of the resource framework in the documentation.