Remove Worker.host_health, update WorkerTable
Previously users would often raise concerns about the total memory
measurement displayed in the workers/ bokeh page, thinking that Dask was
using system memory for data management.  This wasn't the case but we've
removed this value anyway, and the mechanism that was capturing and
reporting system memory in the first place.