Add support for click-based command-line arguments in preload modules.
Adds support for optional command-line parsing to worker and scheduler
--preload modules. Modules *may* support an additional `dask_command`
module attribute, which must provide a click.Command to be used in
parsing command-line options from the root command-line interface.
`dask_command` will be invoked with any "additional" arguments specified
on the command line interface, and is expected to configure the preload
module. Following calls to `dask_setup` and `dask_teardown` provide the
target worker or scheduler objects.