Support tuple-keys in retries
Previously the retries dict passed to the scheduler contained the keys
of the dask collection.  This could be problematic when those keys were
tuples, such as is the case in dask.array and dask.dataframe.  We now
sanitize the keys with the ``tokey`` function.

In the future we might consider cleaning this up ahead of time.
Currently we do it both in retires and in `_graph_to_futures`.