Catch and log OSErrors in diskutils.py
Previously workers would fail when they tried and failed to acquire
file locks in their workspace.  This was due to a variety of issues
ranging from not having permissions (someone else had written such a
file) or the file system not being able to support file-based locks.

Now we allow the system to continue, logging an informative error
along the way.

Fixes #1685
Fixes #1693
Fixes #1707