Allow service ports to be specified as a (host, port) instead
My main use case is to pass `diagnostics_port=(host, port)` to
LocalCluster to run a world-accessible Bokeh server even when the
scheduler uses inproc://. However, the design should be more general
e.g. it is a step towards dask/distributed#1687.

I haven't written any tests or documented it yet - I'd like to get
feedback on whether this is the right approach before investing in
those.