[RFC} Add abstract Cluster class
This defines a superclass and expectations for downstream cluster objects.

This defines expectations both for users and documentation, and also
automated systems like Adaptive.

This also provides a location where we can start to build UI elements
like IPython widgets.

Currently this does not function.  It is only for conversation.