Remove read_into Tornado optimization for Python 2
Python 2 doesn't handle bytearray in the same way as Python 3.
Rather than address this problem we're just going to not support this
optimization for now.  It is only relevant on high bandwidth networks,
and we'd like to encourage people to move to Python 3.

Fixes https://github.com/dask/distributed/issues/1816