Don't penalize subsequent persist calls with lower priority
Previously every call to update_graph would decrement the priority of
those tasks.  This resulted in a first-in-first-out behavior.
This was good both for multi-client workloads, and for consistent
ordering.

However, this got in the way of common workflows where people would
call persist rapidly in sucession:

    x = x.persist()
    y = y.persist()  # <-- these are all lower priority

Now we only decrement priority when we change between clients.

Additionally we re-institute a last-in-first-out policy on workers
that serves only as a tie breaker, such as when using client.submit.