Skip error in get optionally for missing datasets
Include an option to `get` to disable raising a `KeyError` when the
dataset is missing. This helps avoid a fair number of log messages about
the dataset being absent and a `KeyError` being raised in various
contexts. Default to `True` to preserve the existing behavior, but allow
for the `False` option should it be desired.