Protect tcp read in case of connection failure
Previously this would raise in Python 2 with the exception that
frames was undefined in the following line

    msg = yield from_frames(frames, ...)

I don't see a way in which this line should ever have been triggered
and not have frames be well defined.  However, moving this within an
else block on the try-except-else structure resolved the problem.
This avoids errors in Python 2 when workers would sometimes not be
started.  It does not appear to be necessary in Python 3.