Introduce serialization families #1912
This introduces serializers= and deserializers= keywords to Client and
rpc objects, allowing the initiating side of a communication to specify
both the kinds of serializations that it is willing to perform, and also
the kinds of serialization that it is willing to receive.  The other
side of the communication will respect these limitations and select
serialization methods accordingly.

This is useful for a few situations:

1.  Security reasons, for example when people may not want to accept pickled
    results
2.  Asymmetric worker relationships, where some workers may want to
    communicate with others differently, for example when they are on
    the same machine
3.  Inter-language workers where different groups may want to
    communicate with each other using only standardized serialization
    schemes