Use client connection and serialization arguments in progress
Previously the progress bars failed to use client options like
connection_args, serializers, and deserializers when connecting and
communicating with the scheduler.  This resulted in a lack of support
for TLS, and the possibility of dangerous communication if the scheduler
was compromised.

We now capture the client from the passed-in futures and then use these
settings appropriately.

Fixes #2032