Allow worker to refuse data requests with busy signal (#2092)
This allows workers to say "I'm too busy right now" when presented with
a request for data from another worker.  That worker then waits a bit,
queries the scheduler to see if anyone else has that data, and then
tries again.  The wait time is an exponential backoff.

Pragmatically this means that when single pieces of data are in high
demand that the cluster will informally do a tree scattering.  Some workers
will get the data directly while others wait on the busy signal.  Then other
workers will get from them, etc..  We used to ask users to do this explicitly
with the following:

    client.replicate(future)
    or
    client.scatter(data, broadcast=True)

And now the replicate/broadcast step is no longer strictly necessary. (though
some scattering of local data still is).

Machines on the same host are given some preference, and so should be able to
sneak in more easily.