Check self.dependencies when looking at tasks in memory
Previously during a check to see if tasks and their dependencies were already
in memory we assumed that the dependencies were known to the client when it
sent in the computation.  However if the inputs were futures then this was not
the case, and a KeyError was raised because only the scheduler knew where
things were.

Now we properly check both the input dependencies and self.dependencies and are
robust to the information being in either place.

Fixes #2187