Drop the Bokeh index page
As the links on the Bokeh index page are usually incorrect, all the
other apps are accessible from each other, and conventional wisdom is to
direct users to the status page app, the index page is not particularly
useful and can be a bit confusing for the reasons cited. So simply
disable it.