Include worker tls protocol in Scheduler.restart
Fixes https://github.com/dask/distributed/issues/2282

Previously we would try to connect to the Nanny using only the host and
port information, which would default to using the tcp:// protocol and
so dropped all of the security information.  Now we reuse the same
protocol that we use for the worker.