Avoid race condition in Actor's Future
Previously we would err if an actor's future had been created, and the
task finished, but the worker holding the ActorFuture hadn't yet been
informed that the task had finished.  Now we allow the status to be
"pending".

We also move logic from an async test to a sync test because previously
the worker-client would be the same as the main client, which created a
reference cycle and so the future wouldn't get cleaned up.