Fix race condition in normalize_collection (#2386)
This cleans up the use of locks in `Client`, and fixes a race condition
between `Client.normalize_collection` and `Future.release`. We now have
single reentrant-lock on the client, that guards the refcounts on all
futures managed by that client. Any operation that changes this state,
or requires consistency throughout its operation, should acquire this
lock before proceeding. We then guard the critical section of
`normalize_collection` preventing a race condition with
`Future.release`, where previously futures would be decrefed and freed
and then still used in the collection later. A test is added to confirm
the race condition is fixed.