Start fewer but larger workers with LocalCluster
When we have a large number of available cores we should set the default number
of threads per process to be higher than one.

This implements a policy that aims for a number of processes equal to the
square root of the number of cores, at least above a certain amount of cores.

Partially addresses https://github.com/dask/distributed/issues/2450