Find Futures inside SubgraphCallable
New releases of `dask` contain a `SubgraphCallable`, which is an opaque
object that contains a subgraph to be computed as a single task. This
worked fine in many cases, but would fail if that graph contained any
Futures, as distributed wouldn't find them and convert them to data
before computing. We now traverse inside `SubgraphCallable` objects,
pulling out any found futures.

This feels like a bit of a hack to me, but I can't think of a better way
to do this without making more major changes to both codebases.