Remove support for Iterators and Queues
These add non-trivial code complexity, and don't seem to be commonly
used (based on bug reports and SO questions).  They're also a bit odd on
our tests (there are some lingering threads as a result.

This commit removes functionality for them and replaces them with
informative warnings pointing people towards normal for loops.