Remove `Client.upload_environment`
This method was undocumented and broken. Modifying the user environment
is better handled by the deployment solution, the current method is not
resilient to worker additions/removals, and uploading and replacing the
current environment is an inefficient method to add additional packages.
If users really want to upload packages to each worker at runtime, they
can use `upload_file` with a zipfile/egg.