Automatically assign resources based on serialized form of functions
This commit adds functionality to the scheduler to automatically assign
resources to tasks based on terms found in their serialized forms.

This would be useful, for example, to direct all tasks with cupy, cudf,
torch, or tensorflow module names in their serialized form, to be
automatically tagged with a GPU resource tag.  This would simplify
deploying GPU systems considerably, and would also allow for nicer mixed
CPU/GPU workloads.

This should maybe be coupled, in the future, with code that
automatically assigns resources based on the presence of GPUs.