Send noise over the wire to keep connection alive
This change keeps the `dask-ssh` connection alive, preventing ssh
timeouts from closing a pipe and killing workers/schedulers without user
action.

Note that for the `asyncssh` version, this won't be required as
`asyncssh` has a specific `keepalive` parameter that can be passed
through, but before that is the default, this should ease some pain.