Move new SSHCluster to top level
We have two SSHCluster implementations, one that has been around for a
long while, and is still used in the dask-ssh CLI, and a newer
implementation that is based off of SpecCluster.

The old solution is battle hardened, and probably nicer for the CLI, but
not very modifyable.  The new solution is much simpler, and defers most
of the configuration to other systems, notably asyncssh and the
worker/scheduler classes.

This PR keeps both, but moves the new system to the main import, while
gracefully deferring to the old solution if the user seems to be using
those options (it is very obvious which the user intended to use).

We also move SSHCluster to a top-level import so users can do

    from dask.distributed import SSHCluster, Client