Remove circular reference in Cluster
Previously there was a circular reference in
`distributed.deploy.Cluster._watch_worker_status` that would keep the
cluster around after deletion. We fix this here, and add a test. Note
that cleanup isn't instantaneous after `del` since closing down the
cluster takes some time, so the test sleeps in a loop for a bit to check
for this.