Support automatic TLS
This adds support for automatically securing cluster communication with
TLS. This can be useful for situations where you don't have existing
credentials.

This required the following changes:

- ``Security`` objects now support either paths or contents for all
`key`/`cert`/`ca` fields. Due to limitations in Python's ``ssl`` module,
if contents are provided they must be written to a temporary directory
before being loaded back in. We make sure to use secure methods for
doing this. We also change the ``__repr__`` to not show the raw cert
values in the case they're stored in memory.

- ``Security`` objects now have a classmethod ``temporary``, which can
be used to create temporary credentials using self-signed certs. This
requires ``cryptography`` to be installed. Most environments will
already have ``cryptography``, so this isn't a huge new dependency.

```python
>>> sec = Security.temporary()
```

- Both ``Client`` and ``LocalCluster`` now support passing in
``security=True``, which will generate temporary credentials
automatically for use with that cluster. This api could be supported by
other cluster managers, but for now we restrict to ``LocalCluster``
only.

```python
>>> client = Client(security=True)
>>> client
<Client: 'tls://127.0.0.1:52616' processes=4 threads=8, memory=17.18 GB>
```